Refactor Python provider integration: remove PythonProvider class and implement function node retrieval in a new utility
