+++ server/bundled/tools/python/get_function_details.py
@@ -44,7 +44,7 @@ def get_function_details(source_code: str, function_name: str, line_number: int)
         tree = ast.parse(source_code)
         for node in ast.walk(tree):
             if isinstance(node, ast.FunctionDef) and node.name == function_name and node.lineno - 1 == line_number:
-                body_with_lines, raw_body = Analyzer("", source_code, {}, {}).get_function_body(node)
+                body_with_lines, raw_body = Analyzer("", source_code, {}).get_function_body(node)
                 relevant_imports = get_relevant_imports(tree, function_name)
                 function_calls = find_function_calls(tree, function_name)
