+++ server/src/providers/index.ts
@@ -1,4 +1,3 @@
 export * from "./base";
 export * from "./languageProvider";
 export * from "./django/djangoProvider";
-export * from "./python/pythonProvider";
